fix: new structure and style of menu
