+++ src/components/svg/MenuIcon.tsx
@@ -4,7 +4,7 @@ export function OpenMenuIcon() {
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
-      className="fill-slate-100 stroke-white w-7"
+      className="fill-cinder-800 dark:fill-white stroke-cinder-800 dark:stroke-white  w-7"
     >
       <path
         stroke="inherit"
@@ -22,7 +22,7 @@ export function CloseMenuIcon() {
     <svg
       xmlns="http://www.w3.org/2000/svg"
       viewBox="-2 -2 32 32"
-      className="fill-white stroke-white w-6"
+      className="fill-cinder-800 dark:fill-white stroke-cinder-800 dark:stroke-white w-6"
     >
       <path
         fill="inherit"
