+++ src/components/svg/SvgExit.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export const SvgExit = () => {
   return (
     <svg
