+++ src/components/svg/SvgGear.tsx
@@ -1,11 +1,10 @@
-import React from 'react';
 export const SvgGear = () => {
   return (
     <svg
       height="20px"
       width="20px"
       viewBox="0 0 24 24"
-      fill="none"
+      fill="inherit"
       xmlns="http://www.w3.org/2000/svg"
     >
       <g id="SVGRepo_bgCarrier" strokeWidth="0"></g>
