+++ src/hooks/useTheme.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import ThemeContext from 'contexts/themeContext';
 
-export const useTheme = (): string => {
+export const useTheme = (): { theme: string } => {
   const { themeState } = useContext(ThemeContext);
   return { theme: themeState.theme };
 };
